Move HashLayerProof -> fingerprint_strat.rs
